+++ psi/legacy/kmprt17_mp_psi/BUILD.bazel
@@ -29,9 +29,9 @@ psi_cc_library(
         "kmprt17_opprf.h",
     ],
     deps = [
-        "//psi/psi/core:communication",
-        "//psi/psi/utils:sync",
-        "//psi/psi/utils:test_utils",
+        "//psi/utils:communication",
+        "//psi/utils:sync",
+        "//psi/utils:test_utils",
         "@com_google_absl//absl/types:span",
         "@yacl//yacl/base:exception",
         "@yacl//yacl/base:int128",
